Add files via upload
The biggest challenge for any bank is to predict whether or not an applicant is defaulter. Verifying details of each applicant “manually” is the best option, but the challenge here is ‘Number of applications’. Verifying each application manually takes a lot of time and man power.  

This manual process can be automate using “Machine Learning” techniques. Here we are going to predict whether an applicant will be sanctioned loan based on his/her details like marital status, gender, Education, Self-employed, Applicant Income etc.